+++ src/ResourceManager/Compute/Commands.Compute/Extension/AzureVMBackup/SetAzureVMBackupExtension.cs
@@ -75,9 +75,9 @@ namespace Microsoft.Azure.Commands.Compute.Extension.AzureDiskEncryption
             HelpMessage = "The tag for this backup.")]
         public string Tag { get; set; }
 
-        protected override void ProcessRecord()
+        public override void ExecuteCmdlet()
         {
-            base.ProcessRecord();
+            base.ExecuteCmdlet();
 
             AzureVMBackupExtensionUtil azureBackupExtensionUtil = new AzureVMBackupExtensionUtil();