+++ src/ResourceManager/Network/Commands.Network/Models/PSExpressRouteCircuit.cs
@@ -26,6 +26,8 @@ namespace Microsoft.Azure.Commands.Network.Models
 
         public List<PSPeering> Peerings { get; set; }
 
+        public List<PSExpressRouteCircuitAuthorization> Authorizations { get; set; }
+
         public string ServiceKey { get; set; }
 
         public string ServiceProviderNotes { get; set; }
@@ -34,26 +36,30 @@ namespace Microsoft.Azure.Commands.Network.Models
 
         public PSExpressRouteCircuitSku Sku { get; set; }
 
-        public string BillingType { get; set; }
-
         public string ProvisioningState { get; set; }
 
         [JsonIgnore]
         public string SkuText
         {
-            get { return JsonConvert.SerializeObject(Sku, Formatting.Indented); }
+            get { return JsonConvert.SerializeObject(Sku, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
 
         [JsonIgnore]
         public string ServiceProviderPropertiesText
         {
-            get { return JsonConvert.SerializeObject(ServiceProviderProperties, Formatting.Indented); }
+            get { return JsonConvert.SerializeObject(ServiceProviderProperties, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
 
         [JsonIgnore]
         public string PeeringsText
         {
-            get { return JsonConvert.SerializeObject(Peerings, Formatting.Indented); }
+            get { return JsonConvert.SerializeObject(Peerings, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
+        }
+
+        [JsonIgnore]
+        public string AuthorizationsText
+        {
+            get { return JsonConvert.SerializeObject(Authorizations, Formatting.Indented); }
         }
     }
 }