+++ src/ResourceManager/Network/Commands.Network/Models/PSLoadBalancingRule.cs
@@ -32,10 +32,10 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string Protocol { get; set; }
 
         [JsonProperty(Order = 1)]
-        public int FrontendPort { get; set; }
+        public int? FrontendPort { get; set; }
 
         [JsonProperty(Order = 1)]
-        public int BackendPort { get; set; }
+        public int? BackendPort { get; set; }
 
         [JsonProperty(Order = 1)]
         public int? IdleTimeoutInMinutes { get; set; }
@@ -44,7 +44,7 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string LoadDistribution { get; set; }
 
         [JsonProperty(Order = 1)]
-        public bool EnableFloatingIP { get; set; }
+        public bool? EnableFloatingIP { get; set; }
 
         [JsonProperty(Order = 1)]
         public string ProvisioningState { get; set; }
@@ -52,19 +52,19 @@ namespace Microsoft.Azure.Commands.Network.Models
         [JsonIgnore]
         public string FrontendIPConfigurationText
         {
-            get { return JsonConvert.SerializeObject(FrontendIPConfiguration, Formatting.Indented); }
+            get { return JsonConvert.SerializeObject(FrontendIPConfiguration, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
 
         [JsonIgnore]
         public string BackendAddressPoolText
         {
-            get { return JsonConvert.SerializeObject(BackendAddressPool, Formatting.Indented); }
+            get { return JsonConvert.SerializeObject(BackendAddressPool, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
 
         [JsonIgnore]
         public string ProbeText
         {
-            get { return JsonConvert.SerializeObject(Probe, Formatting.Indented); }
+            get { return JsonConvert.SerializeObject(Probe, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
 
     }