+++ src/ResourceManager/Network/Commands.Network/Models/PSNetworkSecurityGroup.cs
@@ -25,34 +25,34 @@ namespace Microsoft.Azure.Commands.Network.Models
 
          public List<PSSecurityRule> DefaultSecurityRules { get; set; }
 
-         public List<PSResourceId> NetworkInterfaces { get; set; }
+         public List<PSNetworkInterface> NetworkInterfaces { get; set; }
 
-         public List<PSResourceId> Subnets { get; set; }
+         public List<PSSubnet> Subnets { get; set; }
 
          public string ProvisioningState { get; set; }
 
          [JsonIgnore]
          public string SecurityRulesText
          {
-             get { return JsonConvert.SerializeObject(SecurityRules, Formatting.Indented); }
+             get { return JsonConvert.SerializeObject(SecurityRules, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
          }
 
          [JsonIgnore]
          public string DefaultSecurityRulesText
          {
-             get { return JsonConvert.SerializeObject(DefaultSecurityRules, Formatting.Indented); }
+             get { return JsonConvert.SerializeObject(DefaultSecurityRules, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
          }
 
          [JsonIgnore]
          public string NetworkInterfacesText
          {
-             get { return JsonConvert.SerializeObject(NetworkInterfaces, Formatting.Indented); }
+             get { return JsonConvert.SerializeObject(NetworkInterfaces, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
          }
 
          [JsonIgnore]
          public string SubnetsText
          {
-             get { return JsonConvert.SerializeObject(Subnets, Formatting.Indented); }
+             get { return JsonConvert.SerializeObject(Subnets, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
          }
      }
 }