+++ src/ResourceManager/NotificationHubs/Commands.NotificationHubs/Commands/Namespace/NewAzureNotificationHubsNamespaceAuthorizationRules.cs
@@ -56,7 +56,7 @@ namespace Microsoft.Azure.Commands.NotificationHubs.Commands.Namespace
         [ValidateNotNullOrEmpty]
         public SharedAccessAuthorizationRuleAttributes SASRule { get; set; }
 
-        protected override void ProcessRecord()
+        public override void ExecuteCmdlet()
         {
             SharedAccessAuthorizationRuleAttributes sasRule = null;
             if (!string.IsNullOrEmpty(InputFile))