+++ src/ResourceManager/TrafficManager/Commands.TrafficManager2/Endpoint/DisableAzureTrafficManagerEndpoint.cs
@@ -47,7 +47,7 @@ namespace Microsoft.Azure.Commands.TrafficManager
         [Parameter(Mandatory = false, HelpMessage = "Do not ask for confirmation.")]
         public SwitchParameter Force { get; set; }
 
-        protected override void ProcessRecord()
+        public override void ExecuteCmdlet()
         {
             var disabled = false;
             TrafficManagerEndpoint endpointToDisable = null;