+++ src/ServiceManagement/ManagedCache/Commands.ManagedCache/Service/SetAzureManagedCacheNamedCache.cs
@@ -53,6 +53,8 @@ namespace Microsoft.Azure.Commands.ManagedCache
 
         public override void ExecuteCmdlet()
         {
+            WriteWarning("Managed Cache will be retired on 11/30/2016. Please migrate to Azure Redis Cache. For more information, see http://go.microsoft.com/fwlink/?LinkID=717458");
+
             if (string.IsNullOrEmpty(ExpiryPolicy))
             {
                 ExpiryPolicy = _DefaultExpiryPolicy;