+++ ChangeLog.md
@@ -1,3 +1,16 @@
+* Azure Redis Cache: new cmdlets for enabling and disabling diagnostics
+  * Set-AzureRmRedisCacheDiagnostics
+  * Remove-AzureRmRedisCacheDiagnostics
+## 2015.12.14 version 1.0.2
+* Azure Compute (ARM):
+  * Enable BGInfo extension by default
+  * Fix the issue when an OS disk is in a different resource group: Now New-AzureRmVM does not create a new storage account for boot diagnostics.
+  * Add Set-AzureRmBginfoExtension cmdlet
+  * Make WinRMCertificateUrl parameter mandatory when Set-AzureRmVMOperatingSystem cmdlet is performed with WinRMHttps switch
+* Azure Compute (Service Management):
+  * Fix the issue when adding a new VM without a data disk
+  * Add ExtensionId parameter for all extension cmdlets
+  * Expose RemoteAccessCertificateThumbprint property for Get-AzureVM cmdlet
 * Azure SQL Database: new cmdlets for managing database threat detection policies:
   * Get-AzureRmSqlDatabaseThreatDetectionPolicy
   * Set-AzureRmSqlDatabaseThreatDetectionPolicy
@@ -5,7 +18,10 @@
 * Azure RemoteApp: New cmdlets for managing stale machine accounts in AD:
   * Get-AzureRemoteAppVmStaleAdObject
   * Clear-AzureRemoteAppVmStaleAdObject
-  
+* ARM Storage:
+  * Fix alias missing issue
+
+
 ## 2015.11.09 version 1.0.1
 * Azure Compute
   * Added cmdlets for managing VM DiskEncryption extension